+++ src/IbanNet/Validation/Methods/LooseValidation.cs
@@ -4,11 +4,14 @@ using IbanNet.Validation.Rules;
 namespace IbanNet.Validation.Methods
 {
 	/// <summary>
-	/// Fast validation consists of all built-in IBAN validation rules, except for checking the entire IBAN structure.
+	/// Loose validation consists of all built-in IBAN validation rules, except for checking the entire IBAN structure.
 	/// This means that the IBAN could potentially contain certain characters that are officially not allowed, but could
 	/// pass all other criteria (even including check digit).
 	/// </summary>
-	public class FastValidation : ValidationMethod
+	/// <remarks>
+	/// Loose validation is around 15%-20% faster than strict validation.
+	/// </remarks>
+	public class LooseValidation : ValidationMethod
 	{
 		internal override IEnumerable<IIbanValidationRule> GetRules()
 		{