+++ src/IbanNet/Validation/Methods/StrictValidation.cs
@@ -6,7 +6,7 @@ namespace IbanNet.Validation.Methods
 	/// <summary>
 	/// Strict validation consists of all built-in IBAN validation rules.
 	/// </summary>
-	public class StrictValidation : FastValidation
+	public class StrictValidation : LooseValidation
 	{
 		internal override IEnumerable<IIbanValidationRule> GetRules()
 		{