+++ test/IbanNet.Tests/LooseIbanValidatorTests.cs
@@ -1,0 +1,14 @@
+﻿using IbanNet.Validation.Methods;
+using NUnit.Framework;
+
+namespace IbanNet
+{
+	[TestFixture]
+	internal class LooseIbanValidatorTests : IbanValidatorIntegrationTests
+	{
+		public LooseIbanValidatorTests()
+			: base(new IbanValidator(new IbanValidatorOptions { ValidationMethod = new LooseValidation() }))
+		{
+		}
+	}
+}